+++ src/hooks/useFillSession.ts
@@ -5,5 +5,6 @@ import { SessionContext } from '../context/sessionContext';
  * @deprecated `useFillActions()` is what you probably want instead
  */
 export function useFillSession() {
-  return useContext(SessionContext);
+  const { session } = useContext(SessionContext);
+  return session;
 }